+++ NEWS.md
@@ -1,3 +1,15 @@
+# Changes in version 0.99.6
+
+* **Breaking change** `read_stockholm_msa()` now returns an S4 object inheriting
+  from class `StockholmMultipleAlignment`. Annotations are stored in slots of
+  class `BStringSet` (GF and GC) or `BStringSetList` (GS and GR), allowing for
+  annotations which do not exist for all sequences.
+* `read_stockholm_msa()` parses GF and GS annotations split over several lines.
+* **Breaking change** `read_stockholm_msa()` can parse amino acid MSA's; the
+  "dna" argument is renamed to "type", and takes the values "RNA", "DNA", and
+  "AA". The default is RNA, as in previous versions.
+* Added `write_stockholm_msa()`.
+
 # Changes in version 0.99.5
 
 * Correct argument `"glocal"` of `cmalign` to `"global"`.  Deprecate `"glocal"`.